Implement better basis set name transformations
There were problems on the website with basis sets containing
slashes in the name (cc-pvXZ(fi/sf/fw), etc). This commit
makes handling of slashes and stars more consistent,
always applying the transformation.

Now, the keys of the metadata dict are transformed names, including
star and slash replacements. You can use either the transformed
or display names when calling the api or using the CLI.

This also required some renaming of test files.