Refactoring in basis set readers
This is a large commit, which does a few things. Mainly, it
completely refactors the basis set readers, adding much
more error checking. It also adds lost of testing
for the readers.

This was done for a few reasons. First, it will catch more
errors in contributed basis sets. This will also allow for more
user-contributed basis sets to be handled in a semi-automated
way.

Second, this is moving towards a more general-purpose conversion
functionality (see issue #76)